tcti: fix pcap timestamp for 32-bit platforms
On most 32-bit platforms, the unix epoch is stored in a 32-bit long
(instead of a 64-bit long) which causes overflows in the pcap
builder and the test code. Add casts to account for 32-bit longs.

+++ src/tss2-tcti/tcti-pcap-builder.c
@@ -290,7 +290,7 @@ pcap_print (
     }
 
     ret = pcap_write_enhanced_packet_block (ctx, buf, pdu_len,
-                                            ts.tv_sec*1000000 + ts.tv_nsec/1000,
+                                            (uint64_t) ts.tv_sec*1000000 + ts.tv_nsec/1000,
                                             payload, payload_len, direction);
     if (ret < 0) {
         goto cleanup;
